refactor namespace helper out to ignoredNamespaces file
