+++ src/lib/assets/ignoredNamespaces.test.ts
@@ -1,9 +1,19 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 import { it, describe, expect, beforeEach } from "vitest";
-import { resolveIgnoreNamespaces } from "./ignoredNamespaces";
+import { resolveIgnoreNamespaces, getIgnoreNamespaces } from "./ignoredNamespaces";
+import { ModuleConfig } from "../types";
+import { WebhookIgnore } from "../k8s";
 
 const originalEnv = { ...process.env };
+
+const ignore = (namespaces?: string[]): WebhookIgnore => ({ namespaces });
+
+const baseConfig = (overrides?: Partial<ModuleConfig>): ModuleConfig => ({
+  uuid: "test-uuid",
+  alwaysIgnore: ignore([]), // default: empty list is fine
+  ...overrides,
+});
 
 beforeEach(() => {
   process.env = { ...originalEnv };
@@ -42,5 +52,58 @@ describe("when namespaces are set in environment variables", () => {
       const result = resolveIgnoreNamespaces();
       expect(result).toEqual(["uds", "project-fox"]);
     });
+  });
+});
+
+describe("getIgnoreNamespaces", () => {
+  it("should use config.alwaysIgnore.namespaces when non-empty", () => {
+    const config = baseConfig({
+      alwaysIgnore: { namespaces: ["ns-a", "ns-b"] },
+      admission: { alwaysIgnore: { namespaces: ["fb-1"] } },
+    });
+
+    const result = getIgnoreNamespaces(config as ModuleConfig);
+
+    expect(result).toEqual(["ns-a", "ns-b"]);
+  });
+
+  it("should use config.admission.alwaysIgnore.namespaces when config.alwaysIgnore.namespaces is empty", () => {
+    const config = baseConfig({
+      alwaysIgnore: { namespaces: [] },
+      admission: { alwaysIgnore: { namespaces: ["fb-a", "fb-b"] } },
+    });
+
+    const result = getIgnoreNamespaces(config as ModuleConfig);
+
+    expect(result).toEqual(["fb-a", "fb-b"]);
+  });
+
+  it("should use config.admission.alwaysIgnore.namespaces when config.alwaysIgnore.namespaces is undefined", () => {
+    const config = baseConfig({
+      // alwaysIgnore present but no namespaces
+      alwaysIgnore: { namespaces: undefined },
+      admission: { alwaysIgnore: { namespaces: ["fb-only"] } },
+    });
+
+    const result = getIgnoreNamespaces(config as ModuleConfig);
+
+    expect(result).toEqual(["fb-only"]);
+  });
+
+  it("should return empty array when both sources are absent/empty", () => {
+    const config = baseConfig({
+      alwaysIgnore: { namespaces: [] },
+      admission: { alwaysIgnore: { namespaces: [] } },
+    });
+
+    const result = getIgnoreNamespaces(config as ModuleConfig);
+
+    expect(result).toEqual([]);
+  });
+
+  it("should return empty array when entire config is undefined", () => {
+    const result = getIgnoreNamespaces(undefined);
+
+    expect(result).toEqual([]);
   });
 });
