+++ src/lib/assets/ignoredNamespaces.ts
@@ -1,5 +1,14 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
+import { ModuleConfig } from "../types";
+
+export function getIgnoreNamespaces(config?: ModuleConfig): string[] {
+  const fromConfig = config?.alwaysIgnore?.namespaces?.length
+    ? config.alwaysIgnore.namespaces
+    : config?.admission?.alwaysIgnore?.namespaces;
+
+  return resolveIgnoreNamespaces(fromConfig);
+}
 
 export function resolveIgnoreNamespaces(ignoredNSConfig: string[] = []): string[] {
   const ignoredNSEnv = process.env.PEPR_ADDITIONAL_IGNORED_NAMESPACES;
